🚧 Start working on split API changes

+++ server/src/models/dto/transactions/new_split_dto.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 #[serde(rename_all = "camelCase")]
 pub struct NewSplitDto<'a> {
     pub description: &'a str,
-    pub amount: i64,
+    pub amount: u32,
     pub category_id: Option<&'a str>,
     pub subcategory_id: Option<&'a str>,
 }
